streamline spec files and update disqus spec to ensure disqus_thread showing
