+++ app/views/shared/_nav.html.erb
@@ -1,5 +1,5 @@
 <nav class="c-nav">
-  <%= link_to "Home", home_path, class: "c-nav__menu-item--link" %>
+  <%= link_to "Home", root_path, class: "c-nav__menu-item--link" %>
   <%= link_to "About", about_path, class: "c-nav__menu-item--link" %>
   <%= link_to "Articles", articles_path, class: "c-nav__menu-item--link" %>
   <%= link_to "Projects", projects_path, class: "c-nav__menu-item--link" %>
