+++ config/routes.rb
@@ -1,7 +1,6 @@
 Rails.application.routes.draw do
   
   root 'static_pages#home'
-  get 'home', to: 'static_pages#home'
   get 'about', to: 'static_pages#about'
   get 'contact', to: 'static_pages#contact'
   get 'projects', to: 'static_pages#projects'
